Move appfooter
