+++ examples/todo/src/ui/TodoApp/MainSection/MainSection.tsx
@@ -1,14 +1,12 @@
 import * as React from 'react'
 import { ToggleAllButton } from './ToggleAllButton'
 import { TodoList } from './TodoList'
-import { AppFooter } from './AppFooter'
 
 export const MainSection = () => {
   return (
     <section className="main">
       <ToggleAllButton />
       <TodoList />
-      <AppFooter />
     </section>
   )
 }
