+++ examples/todo/src/ui/TodoApp/TodoApp.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { AppFooter } from './AppFooter'
 import { HeaderSection } from './HeaderSection'
 import { MainSection } from './MainSection'
 
@@ -7,6 +8,7 @@ export const TodoApp = () => {
     <section className="todoapp">
       <HeaderSection />
       <MainSection />
+      <AppFooter />
     </section>
   )
 }
